chore: move location of tool metrics to a better suited folder
