+++ server/internal/gateway/proxy.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/speakeasy-api/gram/server/internal/templates"
-	tm "github.com/speakeasy-api/gram/server/internal/thirdparty/toolmetrics"
+	tm "github.com/speakeasy-api/gram/server/internal/telemetry"
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/metric"
