+++ server/internal/telemetry/queries.sql.go
@@ -1,4 +1,4 @@
-package toolmetrics
+package telemetry
 
 import (
 	"context"
