+++ server/internal/telemetry/queries_test.go
@@ -1,4 +1,4 @@
-package toolmetrics
+package telemetry
 
 import (
 	"strings"
