+++ server/internal/telemetry/stub.go
@@ -1,4 +1,4 @@
-package toolmetrics
+package telemetry
 
 import (
 	"context"
